chore(badges): centralized used badges throughout
