+++ versioned_docs/version-v6.0.0/dashboard/flows/01_add.mdx
@@ -7,7 +7,7 @@ breadcrumbs: false
 
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
-import { OnlyFolderAdminsBadge, NewSinceBadge } from '@site/src/components/OnlyAdminsBadge';
+import { OnlyFolderAdminsBadge, NewSinceBadge } from '@site/src/components/Badges';
 import { Flows, FlowTraces, FlowTraceImporting } from '@site/src/components/TermHighlight';
 import { TextInput, CheckBoxInput, ToggleInput, DropDownInput, NumberInput, AddInput } from '@site/src/components/InputControls';
 import Breadcrumb from '@site/src/components/Breadcrumb';
