+++ versioned_docs/version-v6.0.0/dashboard/flows/02_search.mdx
@@ -9,7 +9,7 @@ import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import Breadcrumb from '@site/src/components/Breadcrumb';
 import { Flow, FlowTraces } from "@site/src/components/TermHighlight";
-import { NewSinceBadge } from '@site/src/components/OnlyAdminsBadge';
+import { NewSinceBadge } from '@site/src/components/Badges';
 import { faAngleDown, faICursor, faMagnifyingGlass, faArrowPointer } from '@fortawesome/free-solid-svg-icons';
 import { faCircle, faCircleDot, faCalendar, faSquare, faSquareCheck } from '@fortawesome/free-regular-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
