+++ versioned_docs/version-v6.0.0/dashboard/flows/04_import-flow-traces/import flow-via-fa.mdx
@@ -3,7 +3,7 @@ sidebar_label: Import via Function Apps
 title: Import flow traces via Azure Function App logs
 ---
 
-import { NewSinceBadge } from '@site/src/components/OnlyAdminsBadge';
+import { NewSinceBadge } from '@site/src/components/Badges';
 
 # Import flow traces via Azure Function App logs <NewSinceBadge version="v6.2" />
 Invictus allows developers to import flow traces via function app logs coming from Azure Function Apps. These logs translate to the startup of the Function App, but can also contain developer-custom logging. Combined, they result in an execution tree in the Dashboard that represents.
