+++ versioned_docs/version-v6.0.0/dashboard/flows/04_import-flow-traces/import-flow-prog-via-eh.mdx
@@ -2,7 +2,7 @@
 sidebar_label: Import via Event Hubs
 ---
 
-import {OnlyFolderAdminsBadge} from '@site/src/components/OnlyAdminsBadge'
+import {OnlyFolderAdminsBadge} from '@site/src/components/Badges'
 
 # Import flow traces programmatically via Azure Event Hubs
 Invictus allows developers to programmatically import flow traces via an Event Hub. You can locate this resource by the following:
