+++ versioned_docs/version-v6.0.0/dashboard/flows/04_import-flow-traces/import-flows-via-la.mdx
@@ -5,7 +5,7 @@ sidebar_label: Import via Logic Apps
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import { Walkthrough, Collapsible } from '@site/src/components/Walkthrough';
-import { OnlyFolderAdminsBadge, OnlyOperatorsBadge } from '@site/src/components/OnlyAdminsBadge'
+import { OnlyFolderAdminsBadge, OnlyOperatorsBadge } from '@site/src/components/Badges';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {faLink, faSquareCheck, faShield, faAt} from '@fortawesome/free-solid-svg-icons';
 import {faSquare, faCirclePlay} from '@fortawesome/free-regular-svg-icons';
