+++ versioned_docs/version-v6.0.0/dashboard/installation/01_give_ad_access.mdx
@@ -3,7 +3,7 @@ sidebar_label: Give access to your Microsoft Entra ID
 ---
 
 import { Walkthrough, Collapsible } from '@site/src/components/Walkthrough';
-import { OnlyAdminsBadge } from '@site/src/components/OnlyAdminsBadge';
+import { OnlyAdminsBadge } from '@site/src/components/Badges';
 import Breadcrumb from '@site/src/components/Breadcrumb';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {faLink, faSquareCheck, faArrowRight, faChevronRight} from '@fortawesome/free-solid-svg-icons';
