+++ versioned_docs/version-v6.0.0/dashboard/installation/index.mdx
@@ -6,7 +6,7 @@ pagination_next: null
 
 import React from 'react';
 
-import { SharedNote } from '@site/src/components/SharedNoteBadge';
+import { SharedNote } from '@site/src/components/Badges';
 import { Walkthrough, Step } from '@site/src/components/Walkthrough';
 
 import ParameterTable from "@site/src/components/ParameterTable";
